+++ lib/layout/constants/sizes.constant.dart
@@ -8,9 +8,9 @@ class CLSizes {
   static const xl = 32.0;
   static const xxl = 40.0;
   static const xxxl = 48.0;
-  static const borderRadius = 12.0;
-  static const radiusSm = 8.0;
-  static const radiusLg = 16.0;
+  static const borderRadius = 6.0;
+  static const radiusSm = 4.0;
+  static const radiusLg = 8.0;
   static const padding = 18.0;
   static const opacity = 0.2;
   static const verticalPadding = 16.0;